fix: Use -pie instead of -static on linking
- Add -Wall -Wextra -Werror to CFLAGS
- Fix some assebly files to make the code position independent (PIE/PIC)
- Fix test.c

+++ test.c
@@ -58,7 +58,7 @@ int main(void)
 			const char *message = "libasm";
 			const int fd = open("/dev/null", O_WRONLY, 0);
 			const ssize_t ret = ft_write(fd, message, strlen(message));
-			assert(ret == strlen(message));
+			assert(ret == (ssize_t) strlen(message));
 			close(fd);
 		}
